Crate spl_tlv_account_resolution

Crate defining a state interface for offchain account resolution. If a program writes the proper state information into one of their accounts, any offchain and onchain client can fetch any additional required accounts for an instruction.

Re-exports

pub use solana_program;

Modules

account

Struct for managing extra required account configs, ie. defining accounts required for your interface program, which can be AccountMetas - which have fixed addresses - or PDAs - which have addresses derived from a collection of seeds

error

Error types

seeds

Types for managing seed configurations in TLV Account Resolution

state

State transition types